# bedelia-rest/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bedelia UTEC",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/administrativos/{ci}/sede/": {
            "get": {
                "tags": [
                    "Administrativos"
                ],
                "description": "Devuelve la sede en la que trabaja el usuario administrativo",
                "operationId": "App\\Http\\Controllers\\AdministrativosController::obtenerSede",
                "parameters": [
                    {
                        "name": "ci",
                        "in": "path",
                        "description": "Cédula del usuario administrativo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Devuelve la sede donde trabaja el usuario administratvo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SedeDTO"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Administrativos"
                ],
                "description": "Establece la sede en la que trabaja el usuario administrativo",
                "operationId": "App\\Http\\Controllers\\AdministrativosController::establecerSede",
                "parameters": [
                    {
                        "name": "ci",
                        "in": "path",
                        "description": "Cédula del usuario administrativo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SedeDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Devuelve la sede donde trabaja el usuario administratvo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SedeDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DireccionDTO": {
                "properties": {
                    "departamento": {
                        "type": "string"
                    },
                    "ciudad": {
                        "type": "string"
                    },
                    "calle": {
                        "type": "string"
                    },
                    "numero": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "SedeDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "telefono": {
                        "type": "string"
                    },
                    "direccion": {
                        "$ref": "#/components/schemas/DireccionDTO"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "Token de autenticacion conformato como '<token>'",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}